+++ src/LionWeb-CSharp/core/VersionSpecific/V2023_1/BuiltInsLanguage_2023_1.cs
@@ -15,59 +15,71 @@
 // SPDX-FileCopyrightText: 2024 TRUMPF Laser SE and other contributors
 // SPDX-License-Identifier: Apache-2.0
 
-namespace LionWeb.Core.M2;
+// ReSharper disable InconsistentNaming
+namespace LionWeb.Core.VersionSpecific.V2023_1;
 
+using M2;
 using M3;
 
 /// <inheritdoc />
-public interface IBuiltInsLanguage_2024_1 : IBuiltInsLanguage;
+public interface IBuiltInsLanguage_2023_1 : IBuiltInsLanguage
+{
+    /// <summary>
+    /// The built-in primitive type Json.
+    /// </summary>
+    PrimitiveType Json { get; }
+}
 
-/// <inheritdoc cref="IBuiltInsLanguage_2024_1" />
-public sealed class BuiltInsLanguage_2024_1 :
-    LanguageBase<BuiltInsFactory_2024_1>, IBuiltInsLanguage_2024_1
+/// <inheritdoc cref="IBuiltInsLanguage_2023_1" />
+public sealed class BuiltInsLanguage_2023_1 : LanguageBase<BuiltInsFactory_2023_1>, IBuiltInsLanguage_2023_1 
 {
     /// <summary>
     /// The definition of the LionCore built-ins language.
     /// </summary>
-    public static readonly BuiltInsLanguage_2024_1 Instance = new Lazy<BuiltInsLanguage_2024_1>(() => new()).Value;
+    public static readonly BuiltInsLanguage_2023_1 Instance = new Lazy<BuiltInsLanguage_2023_1>(() => new()).Value;
 
     private const string _name = "LionCore_builtins";
 
     /// <inheritdoc />
-    protected override IBuiltInsLanguage_2024_1 _builtIns => this;
+    protected override IBuiltInsLanguage_2023_1 _builtIns => this;
 
     /// <inheritdoc />
-    protected override ILionCoreLanguage_2024_1 _m3 => LionCoreLanguage_2024_1.Instance;
+    protected override ILionCoreLanguage_2023_1 _m3 => LionCoreLanguage_2023_1.Instance;
 
-    internal BuiltInsLanguage_2024_1() : base(LionCoreBuiltInsIdAndKey, LionWebVersions.v2024_1)
+    internal BuiltInsLanguage_2023_1() : base(LionCoreBuiltInsIdAndKey, LionWebVersions.v2023_1)
     {
-        _boolean = new(() => new PrimitiveTypeBase<BuiltInsLanguage_2024_1>($"{LionCoreBuiltInsIdAndKey}-Boolean", this)
+        _boolean = new(() => new PrimitiveTypeBase<BuiltInsLanguage_2023_1>($"{LionCoreBuiltInsIdAndKey}-Boolean", this)
         {
             Key = $"{LionCoreBuiltInsIdAndKey}-Boolean", Name = "Boolean"
         });
 
-        _integer = new(() => new PrimitiveTypeBase<BuiltInsLanguage_2024_1>($"{LionCoreBuiltInsIdAndKey}-Integer", this)
+        _integer = new(() => new PrimitiveTypeBase<BuiltInsLanguage_2023_1>($"{LionCoreBuiltInsIdAndKey}-Integer", this)
         {
             Key = $"{LionCoreBuiltInsIdAndKey}-Integer", Name = "Integer"
         });
 
-        _string = new(() => new PrimitiveTypeBase<BuiltInsLanguage_2024_1>($"{LionCoreBuiltInsIdAndKey}-String", this)
+        _json = new(() => new PrimitiveTypeBase<BuiltInsLanguage_2023_1>($"{LionCoreBuiltInsIdAndKey}-JSON", this)
+        {
+            Key = $"{LionCoreBuiltInsIdAndKey}-JSON", Name = "JSON"
+        });
+
+        _string = new(() => new PrimitiveTypeBase<BuiltInsLanguage_2023_1>($"{LionCoreBuiltInsIdAndKey}-String", this)
         {
             Key = $"{LionCoreBuiltInsIdAndKey}-String", Name = "String"
         });
 
         _iNamed = new(() =>
-            new InterfaceBase<BuiltInsLanguage_2024_1>($"{LionCoreBuiltInsIdAndKey}-INamed", this)
+            new InterfaceBase<BuiltInsLanguage_2023_1>($"{LionCoreBuiltInsIdAndKey}-INamed", this)
             {
                 Key = $"{LionCoreBuiltInsIdAndKey}-INamed", Name = "INamed", FeaturesLazy = new(() => [INamed_name])
             });
         _iNamed_name = new(() =>
-            new PropertyBase<BuiltInsLanguage_2024_1>($"{LionCoreBuiltInsIdAndKey}-INamed-name", INamed, this)
+            new PropertyBase<BuiltInsLanguage_2023_1>($"{LionCoreBuiltInsIdAndKey}-INamed-name", INamed, this)
             {
                 Key = $"{LionCoreBuiltInsIdAndKey}-INamed-name", Name = "name", Optional = false, Type = String
             });
 
-        _node = new(() => new ConceptBase<BuiltInsLanguage_2024_1>($"{LionCoreBuiltInsIdAndKey}-Node", this)
+        _node = new(() => new ConceptBase<BuiltInsLanguage_2023_1>($"{LionCoreBuiltInsIdAndKey}-Node", this)
         {
             Key = $"{LionCoreBuiltInsIdAndKey}-Node", Name = "Node", Abstract = true, Partition = false
         });
@@ -89,6 +101,7 @@ public sealed class BuiltInsLanguage_2024_1 :
         [
             Boolean,
             Integer,
+            Json,
             String,
             INamed,
             Node
@@ -99,7 +112,7 @@ public sealed class BuiltInsLanguage_2024_1 :
     public override IReadOnlyList<Language> DependsOn { get => []; }
 
     /// <inheritdoc />
-    public override BuiltInsFactory_2024_1 GetFactory() => new(this);
+    public override BuiltInsFactory_2023_1 GetFactory() => new(this);
 
     private readonly Lazy<PrimitiveType> _boolean;
 
@@ -110,6 +123,11 @@ public sealed class BuiltInsLanguage_2024_1 :
 
     /// <inheritdoc />
     public PrimitiveType Integer => _integer.Value;
+
+    private readonly Lazy<PrimitiveType> _json;
+
+    /// <inheritdoc />
+    public PrimitiveType Json => _json.Value;
 
     private readonly Lazy<PrimitiveType> _string;
 
@@ -137,12 +155,12 @@ public sealed class BuiltInsLanguage_2024_1 :
     public const string LionCoreBuiltInsIdAndKey = "LionCore-builtins";
 }
 
-/// Factory for <see cref="BuiltInsLanguage_2024_1"/>.
-public sealed class BuiltInsFactory_2024_1 : INodeFactory
+/// Factory for <see cref="BuiltInsLanguage_2023_1"/>.
+public sealed class BuiltInsFactory_2023_1 : INodeFactory
 {
-    private readonly BuiltInsLanguage_2024_1 _language;
+    private readonly BuiltInsLanguage_2023_1 _language;
 
-    internal BuiltInsFactory_2024_1(BuiltInsLanguage_2024_1 language)
+    internal BuiltInsFactory_2023_1(BuiltInsLanguage_2023_1 language)
     {
         _language = language;
     }
